Fixed test code

+++ src/Test/L0/Worker/TestResults/Legacy/CtestResultReaderTests.cs
@@ -244,13 +244,20 @@ namespace Microsoft.VisualStudio.Services.Agent.Tests.Worker.Results
 
         public void Dispose()
         {
-            try
+            Dispose(true);
+        }
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
             {
-                File.Delete(_fileName);
-            }
-            catch
-            {
+                try
+                {
+                    File.Delete(_fileName);
+                }
+                catch
+                {
 
+                }
             }
         }
