+++ src/Test/L0/Worker/TestResults/Legacy/JunitResultReaderTests.cs
@@ -623,14 +623,21 @@ namespace Microsoft.VisualStudio.Services.Agent.Tests.Worker.TestResults
 
         public void Dispose()
         {
-            _jUnitReader.AddResultsFileToRunLevelAttachments = true;
-            try
+            Dispose(true);
+        }
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
             {
-                File.Delete(_fileName);
-            }
-            catch
-            {
+                _jUnitReader.AddResultsFileToRunLevelAttachments = true;
+                try
+                {
+                    File.Delete(_fileName);
+                }
+                catch
+                {
 
+                }
             }
         }
     }
